fix: naming conventions of location of hooks by moving hooks to hooks folder
