+++ packages/ui/components/index.tsx
@@ -3,4 +3,3 @@ export * from './ErrorView';
 export * from './Layout';
 export * from './LoadingView';
 export * from './Navbar';
-export * from './usePluginProvider';
